Rationalize different Extension typings
Extension and GeminiGLIExtension had drifted and there were references
to the internal file type of the extension json file. Collapsed the two
typings into one and ensured the internal file structure was not
accessed outside of loading the extension.

This is a preparation for being able to keep all extensions loaded but
just marked as active or inactive (which will then enable dynamic
activation)