fix(cli): Use correct defaults for file filtering
File filtering options should use the defaults when they are not
overridden in settings files. This fixes a bug where the defaults for
memory files were used instead which led to `.gitignore` filtering being
disabled unless explicitly overridden.

A test is added to reproduce the issue and prevent further regressions.