refactor: Switch over to unified shouldIgnoreFile
Converts all legacy usage of the `shouldGitIgnoreFile` and
`shouldGeminiIgnoreFile` helpers to the unified `shouldIgnoreFile`
function.

This both simplifies code and makes it more correct by respecting the
user's ignore file configuration. The legacy functions are now unused
and removed. Internally `shouldIgnoreFile` is converted over to
calling `filterFiles` to remove its dependency on the legacy helpers and
avoid code duplication.