fix(cli): cancel ongoing request on `esc` iff in NORMAL with vim mode
When vim mode is toggled on, cancel ongoing requests if and only if the
text buffer is in NORMAL mode; letting the vim context automatically
switch to NORMAL mode otherwise.

This ensures that we can still interact with the text buffer in normal
mode even when a request is ongoing.