feat(tables): reduce the number of columns subject to overflow.
The old algorithm for calculating column widths when there is table overflow simply scaled every column width proportionally. Because it had to round down,  this resulted in at least one cell in each column experiencing overflow. This resulted in unnecessary trunactions, or with this PR would result in unnecessary wrapping and increases in table height.

The new algorithm tries to evenly, rather than proportionally, distribute the available width to columns. This results in fewer columns having overflowing cells, while dedicating approximately the same width to each overflowing column.

I also made a small tweak to ensure that an overflowing table always uses 100% of the available width.